+++ projectroles/forms.py
@@ -256,9 +256,9 @@ class ProjectForm(SODARModelForm):
                 roles__in=RoleAssignment.objects.filter(
                     user=user,
                     role__name__in=[
-                        'project owner',
-                        'project delegate',
-                        'project contributor',
+                        PROJECT_ROLE_OWNER,
+                        PROJECT_ROLE_DELEGATE,
+                        PROJECT_ROLE_CONTRIBUTOR,
                     ],
                 )
             )
@@ -310,7 +310,7 @@ class ProjectForm(SODARModelForm):
                     APP_SETTING_SCOPE_PROJECT, plugin=plugin, **self.p_kwargs
                 )
             else:
-                name = 'projectroles'
+                name = APP_NAME
                 p_settings = self.app_settings.get_setting_defs(
                     APP_SETTING_SCOPE_PROJECT, app_name=name, **self.p_kwargs
                 )
@@ -346,7 +346,6 @@ class ProjectForm(SODARModelForm):
                                 project=self.instance,
                             )
                         )
-
                     else:
                         self.initial[s_field] = json.dumps(
                             self.app_settings.get_default_setting(
@@ -441,7 +440,6 @@ class ProjectForm(SODARModelForm):
 
         # Access parent project if present
         parent_project = None
-
         if project:
             parent_project = Project.objects.filter(sodar_uuid=project).first()
 
@@ -464,7 +462,6 @@ class ProjectForm(SODARModelForm):
 
         # Modify ModelChoiceFields to use sodar_uuid
         self.fields['parent'].to_field_name = 'sodar_uuid'
-
         # Set readme widget with preview
         self.fields['readme'].widget = PagedownWidget(
             attrs={'show_preview': True}
@@ -474,16 +471,12 @@ class ProjectForm(SODARModelForm):
         if self.instance.pk:
             # Set readme value as raw markdown
             self.initial['readme'] = self.instance.readme.raw
-
             # Hide project type selection
             self.fields['type'].widget = forms.HiddenInput()
-
             # Set hidden project field for autocomplete
             self.initial['project'] = self.instance
-
             # Set owner value
             self.initial['owner'] = self.instance.get_owner().user.sodar_uuid
-
             # Hide owner widget if updating (changed in member modification UI)
             self.fields['owner'].widget = forms.HiddenInput()
 
@@ -509,7 +502,6 @@ class ProjectForm(SODARModelForm):
         else:
             # Set hidden project field for autocomplete
             self.initial['project'] = None
-
             # Hide parent selection
             self.fields['parent'].widget = forms.HiddenInput()
 
@@ -517,9 +509,8 @@ class ProjectForm(SODARModelForm):
             if parent_project:
                 # Parent must be current parent
                 self.initial['parent'] = parent_project.sodar_uuid
-
-                # Set current user as initial value
-                self.initial['owner'] = self.current_user.sodar_uuid
+                # Set owner of parent category as initial owner
+                self.initial['owner'] = parent_project.get_owner().user
 
                 # If current user is not parent owner, disable owner select
                 if (
@@ -533,15 +524,14 @@ class ProjectForm(SODARModelForm):
                 # Force project type
                 if settings.PROJECTROLES_DISABLE_CATEGORIES:
                     self.initial['type'] = PROJECT_TYPE_PROJECT
-
                 else:
                     self.initial['type'] = PROJECT_TYPE_CATEGORY
-
                 # Hide project type selection
                 self.fields['type'].widget = forms.HiddenInput()
-
                 # Set up parent field
                 self.initial['parent'] = None
+                # Set current user as initial owner
+                self.initial['owner'] = self.current_user
 
         if self.instance.is_remote():
             self.fields['title'].widget = forms.HiddenInput()
@@ -566,7 +556,6 @@ class ProjectForm(SODARModelForm):
                     'parent',
                     'You do not have permission to place a category under root',
                 )
-
             elif (
                 self.cleaned_data.get('type') == PROJECT_TYPE_PROJECT
                 and not settings.PROJECTROLES_DISABLE_CATEGORIES
@@ -643,7 +632,6 @@ class ProjectForm(SODARModelForm):
                         self.cleaned_data[s_field] = json.loads(
                             self.cleaned_data.get(s_field)
                         )
-
                     except json.JSONDecodeError as err:
                         # TODO: Shouldn't we use add_error() instead?
                         raise forms.ValidationError(